rename SearchAlgolia -> SearchBar
