+++ packages/docusaurus-theme-classic/src/theme/SearchBar.js
