+++ packages/docusaurus-theme-search-algolia/src/theme/SearchBar/index.js
