+++ packages/docusaurus-theme-search-algolia/src/theme/SearchBar/styles.css
